fix guard buy price to float64

+++ handlers/guard_buy.go
@@ -60,7 +60,7 @@ func guardBuyMsg(data *blive.LiveData) error {
 		TargetUid: targetUid,
 		Command:   blive.GuardBuy,
 		Display:   display,
-		Price:     guardBuy.Price,
+		Price:     float64(guardBuy.Price),
 	}
 
 	result := db.Database.Create(behaviour)
